requirements.txt
#python
# Python 3.9.6


accelerate==0.29.3
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
environs==9.5.0
eval_type_backport==0.2.0
exceptiongroup==1.2.1
filelock==3.15.4
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
llama-cloud==0.0.9
llama-index==0.10.56
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.12
llama-index-core==0.10.56
llama-index-embeddings-openai==0.1.11
llama-index-embeddings-together==0.1.3
llama-index-indices-managed-llama-cloud==0.2.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.26
llama-index-llms-openai-like==0.1.3
llama-index-llms-together==0.1.3
llama-index-multi-modal-llms-openai==0.1.8
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.30
llama-index-readers-llama-parse==0.1.6
llama-index-vector-stores-milvus==0.1.21
llama-index-vector-stores-postgres==0.1.12
llama-parse==0.4.9
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
milvus-lite==2.4.8
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.35.10
packaging==24.1
pandas==2.2.2
pgvector==0.2.5
pillow==10.4.0
protobuf==5.27.2
psutil==6.0.0
psycopg2-binary==2.9.9
pyarrow==16.1.0
PyAudio==0.2.14
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pymilvus==2.4.4
pypdf==4.3.0
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scipy==1.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.31
striprtf==0.0.26
sympy==1.12.1
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
together==1.2.1
tokenizers==0.19.1
torch==2.0.1
tqdm==4.66.4
transformers==4.40.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2